requirements.txt
# AutoGen multi-agent framework
pyautogen>=0.2,<0.3

# Config
python-dotenv>=1.0.0
